Refactored to avoid overwriting models
