+++ auth.js
@@ -1,12 +1,9 @@
 const crypto_js = require('crypto-js')
 const UIDGenerator = require('uid-generator')
 const uidgen = new UIDGenerator();
-const { author_scheme, login_scheme } = require('./db_schema/author_schema.js');
+const { Author, Login } = require('./db_schema/author_schema.js');
 const mongoose = require('mongoose');
 mongoose.set('strictQuery', true);
-const database = mongoose.connection;
-const Login = database.model('Login', login_scheme);
-const Author = database.model('Author', author_scheme);
 
 async function checkUsername(req) {
     const author = await Author.findOne({username: req.body.username});
